+++ failfast-examples/src/test/kotlin/failgood/examples/AllTests.kt
@@ -1,9 +1,0 @@
-package failgood.examples
-
-import failgood.FailFast
-
-
-fun main() {
-    FailFast.runAllTests()
-}
-