+++ failfast-examples/src/test/kotlin/failgood/examples/AutoTestMain.kt
@@ -1,8 +1,0 @@
-package failgood.examples
-
-import failgood.FailFast.autoTest
-
-
-fun main() {
-    autoTest()
-}