+++ failfast-junit-engine-it/src/test/kotlin/failgood/junit/it/AllTests.kt
@@ -1,9 +1,0 @@
-package failgood.junit.it
-
-import failgood.FailFast
-
-
-fun main() {
-    FailFast.runAllTests()
-}
-