+++ failfast-junit-engine-it/src/test/kotlin/failgood/junit/it/AutoTestMain.kt
@@ -1,8 +1,0 @@
-package failgood.junit.it
-
-import failgood.FailFast.autoTest
-
-
-fun main() {
-    autoTest()
-}