+++ failfast/src/main/kotlin/failgood/FailFastDSL.kt
@@ -1,4 +1,0 @@
-package failgood
-
-@DslMarker
-annotation class FailFastDSL