+++ failfast/src/main/kotlin/failgood/TestDSL.kt
@@ -1,7 +1,0 @@
-package failgood
-
-@FailFastDSL
-interface TestDSL : ResourcesDSL {
-    suspend fun println(body: String)
-}
-