+++ failfast/src/test/kotlin/failgood/AutoTestMain.kt
@@ -1,7 +1,0 @@
-package failgood
-
-import failgood.FailFast.autoTest
-
-fun main() {
-    autoTest()
-}