+++ failgood-examples/src/test/kotlin/failgood/examples/AllTests.kt
@@ -1,0 +1,9 @@
+package failgood.examples
+
+import failgood.FailGood
+
+
+fun main() {
+    FailGood.runAllTests()
+}
+