+++ failgood-junit-engine-it/src/test/kotlin/failgood/junit/it/AllTests.kt
@@ -1,0 +1,9 @@
+package failgood.junit.it
+
+import failgood.FailGood
+
+
+fun main() {
+    FailGood.runAllTests()
+}
+