+++ failgood-junit-engine-it/src/test/kotlin/failgood/junit/it/AutoTestMain.kt
@@ -1,0 +1,8 @@
+package failgood.junit.it
+
+import failgood.FailGood.autoTest
+
+
+fun main() {
+    autoTest()
+}