+++ failgood/src/main/kotlin/failgood/TestDSL.kt
@@ -1,0 +1,7 @@
+package failgood
+
+@FailGoodDSL
+interface TestDSL : ResourcesDSL {
+    suspend fun println(body: String)
+}
+