+++ failgood/src/test/kotlin/failgood/AutoTestMain.kt
@@ -1,0 +1,7 @@
+package failgood
+
+import failgood.FailGood.autoTest
+
+fun main() {
+    autoTest()
+}